requirements.txt
gunicorn==19.4.5
Django==2.0
libsass==0.13.6
psycopg2==2.7.3.2
django-compressor
django-sass-processor==0.5.7
easy-thumbnails==2.5.0
django-image-cropping==1.1.0
django-ordered-model==1.4.3
django-ckeditor==5.4.0
django-solo==1.1.3
django-phonenumber-field==2.0.0
